Fix: make every single-precision promotion and narrowing in the float tier explicit.
- Add scripts/float_tier_diagnostics.sh, which compiles the three single-precision translation units under g++ and clang with the project warning set and counts the double-promotion and float-conversion diagnostics per unit and per compiler.
- Spell the seven inexact decimal defaults as explicit conversions instead of braced ones, in the particle filter, the multiplicative and manifold attitude filters, the complementary filter, the sigma-point options, the rotation logarithm and the compile witness. Every value is bit-identical to what it was, at float and at double.
- Cast the matched value, the target and the tolerance to the test framework's own double parameter type at the three float-tier assertion sites, so the promotion the matcher's signature already performed is written down rather than implied. No tolerance and no expected value moved.
- Construct the particle filter, the manifold filter and the double-S trajectory at single precision in the instantiation tier, which previously covered neither, and move that unit from the exceptions carve-out into the default tree, where it builds and passes throw-free.
- Realign the sigma-point options block quoted in the unscented filter documentation with the header.

+++ tests/unit/float_tier_test.cpp
@@ -63,8 +63,13 @@ TEST_CASE("place() accepts a numerically-conjugate float pole pair the way it do
 
     const float eps = std::numeric_limits<float>::epsilon();
     const float tol = 2.0f * eps * (1.0f + static_cast<float>(K_ref->norm()));
-    REQUIRE_THAT((*K_f)(0, 0), WithinAbs(static_cast<float>((*K_ref)(0, 0)), tol));
-    REQUIRE_THAT((*K_f)(0, 1), WithinAbs(static_cast<float>((*K_ref)(0, 1)), tol));
+
+    // WithinAbs is declared over double, so it is the matcher's signature and
+    // not the library that promotes; the casts spell that promotion out.
+    REQUIRE_THAT(static_cast<double>((*K_f)(0, 0)),
+                 WithinAbs(static_cast<double>(static_cast<float>((*K_ref)(0, 0))), static_cast<double>(tol)));
+    REQUIRE_THAT(static_cast<double>((*K_f)(0, 1)),
+                 WithinAbs(static_cast<double>(static_cast<float>((*K_ref)(0, 1))), static_cast<double>(tol)));
 }
 
 TEST_CASE("biquad steady-state reset() at float matches the double reference near a singular DC gain", "[float][anchor]")
@@ -98,5 +103,6 @@ TEST_CASE("biquad steady-state reset() at float matches the double reference nea
 
     const float eps = std::numeric_limits<float>::epsilon();
     const float tol = 2.0f * eps * (1.0f + std::abs(static_cast<float>(y_d)));
-    REQUIRE_THAT(y_f, WithinAbs(static_cast<float>(y_d), tol));
+    REQUIRE_THAT(static_cast<double>(y_f),
+                 WithinAbs(static_cast<double>(static_cast<float>(y_d)), static_cast<double>(tol)));
 }
